+++ OctoAwesome/OctoAwesome.Client/Screens/OptionsScreen.cs
@@ -10,24 +10,17 @@ using System.Linq;
 
 namespace OctoAwesome.Client.Screens
 {
-    class OptionsScreen : BaseScreen
+    internal sealed class OptionsScreen : BaseScreen
     {
-        private OctoGame game;
         private AssetComponent assets;
 
         private Button exitButton;
-        private Label rangeTitle;
-        private Textbox mapPath;
-
-        private ISettings settings;
 
         public OptionsScreen(ScreenComponent manager) : base(manager)
         {
-            game = manager.Game;
-            assets = game.Assets;
+            assets = manager.Game.Assets;
 
             Padding = new Border(0, 0, 0, 0);
-            settings = manager.Game.Settings;
 
             Title = Languages.OctoClient.Options;
 
@@ -50,147 +43,12 @@ namespace OctoAwesome.Client.Screens
             TabPage optionsPage = new TabPage(manager, Languages.OctoClient.Options);
             tabs.Pages.Add(optionsPage);
 
-            ////////////////////////////////////////////Settings Stack////////////////////////////////////////////
-            StackPanel settingsStack = new StackPanel(manager)
-            {
-                Orientation = Orientation.Vertical,
-                VerticalAlignment = VerticalAlignment.Top,
-                Padding = new Border(20, 20, 20, 20),
-                Width = 650
-            };
-            optionsPage.Controls.Add(settingsStack);
-
-            //////////////////////Viewrange//////////////////////
-            string viewrange = settings.Get<string>("Viewrange");
-
-            rangeTitle = new Label(manager)
-            {
-                Text = Languages.OctoClient.Viewrange + ": " + viewrange
-            };
-            settingsStack.Controls.Add(rangeTitle);
-
-            Slider viewrangeSlider = new Slider(manager)
+            OptionsOptionControl optionsOptions = new OptionsOptionControl(manager, this)
             {
                 HorizontalAlignment = HorizontalAlignment.Stretch,
-                Height = 20,
-                Range = 9,
-                Value = int.Parse(viewrange) - 1
+                VerticalAlignment = VerticalAlignment.Stretch,
             };
-            viewrangeSlider.ValueChanged += (value) => SetViewrange(value + 1);
-            settingsStack.Controls.Add(viewrangeSlider);
-
-
-            //////////////////////Persistence//////////////////////
-            StackPanel persistenceStack = new StackPanel(manager)
-            {
-                Orientation = Orientation.Horizontal,
-                Margin = new Border(0, 20, 0, 0)
-            };
-            settingsStack.Controls.Add(persistenceStack);
-
-            Label persistenceTitle = new Label(manager)
-            {
-                Text = Languages.OctoClient.DisablePersistence + ":"
-            };
-            persistenceStack.Controls.Add(persistenceTitle);
-
-            Checkbox disablePersistence = new Checkbox(manager)
-            {
-                Checked = bool.Parse(settings.Get<string>("DisablePersistence")),
-                HookBrush = new TextureBrush(assets.LoadTexture(typeof(ScreenComponent), "iconCheck_brown"), TextureBrushMode.Stretch),
-            };
-            disablePersistence.CheckedChanged += (state) => SetPersistence(state);
-            persistenceStack.Controls.Add(disablePersistence);
-
-            //////////////////////Map Path//////////////////////
-            StackPanel mapPathStack = new StackPanel(manager)
-            {
-                Orientation = Orientation.Vertical,
-                Margin = new Border(0, 20, 0, 0),
-                HorizontalAlignment = HorizontalAlignment.Stretch
-            };
-            settingsStack.Controls.Add(mapPathStack);
-
-            mapPath = new Textbox(manager)
-            {
-                Text = settings.Get<string>("ChunkRoot"),
-                Enabled = false,
-                HorizontalAlignment = HorizontalAlignment.Stretch,
-                Background = new BorderBrush(Color.LightGray, LineType.Solid, Color.Gray)
-            };
-            mapPathStack.Controls.Add(mapPath);
-
-            Button changePath = Button.TextButton(manager, Languages.OctoClient.ChangePath);
-            changePath.HorizontalAlignment = HorizontalAlignment.Center;
-            changePath.Height = 40;
-            changePath.LeftMouseClick += (s, e) => ChangePath();
-            mapPathStack.Controls.Add(changePath);
-
-            //////////////////////Fullscreen//////////////////////
-            StackPanel fullscreenStack = new StackPanel(manager)
-            {
-                Orientation = Orientation.Horizontal,
-                Margin = new Border(0, 20, 0, 0)
-            };
-            settingsStack.Controls.Add(fullscreenStack);
-
-            Label fullscreenTitle = new Label(manager)
-            {
-                Text = Languages.OctoClient.EnableFullscreenOnStartup + ":"
-            };
-            fullscreenStack.Controls.Add(fullscreenTitle);
-
-            Checkbox enableFullscreen = new Checkbox(manager)
-            {
-                Checked = bool.Parse(settings.Get<string>("EnableFullscreen")),
-                HookBrush = new TextureBrush(assets.LoadTexture(typeof(ScreenComponent), "iconCheck_brown"), TextureBrushMode.Stretch),
-            };
-            enableFullscreen.CheckedChanged += (state) => SetFullscreen(state);
-            fullscreenStack.Controls.Add(enableFullscreen);
-
-            //////////////////////Auflösung//////////////////////
-            StackPanel resolutionStack = new StackPanel(manager)
-            {
-                Orientation = Orientation.Horizontal,
-                Margin = new Border(0, 20, 0, 0)
-            };
-            settingsStack.Controls.Add(resolutionStack);
-
-            Label resolutionTitle = new Label(manager)
-            {
-                Text = Languages.OctoClient.Resolution + ":"
-            };
-            resolutionStack.Controls.Add(resolutionTitle);
-
-            Textbox resolutionWidthTextbox = new Textbox(manager)
-            {
-                Text = settings.Get<string>("Width"),
-                Width = 50,
-                Background = new BorderBrush(Color.LightGray, LineType.Solid, Color.Gray)
-            };
-            resolutionWidthTextbox.TextChanged += ResolutionWidthTextbox_TextChanged;
-            resolutionStack.Controls.Add(resolutionWidthTextbox);
-
-            Label xLabel = new Label(manager)
-            {
-                Text = "x"
-            };
-            resolutionStack.Controls.Add(xLabel);
-
-            Textbox resolutionHeightTextbox = new Textbox(manager)
-            {
-                Text = settings.Get<string>("Height"),
-                Width = 50,
-                Background = new BorderBrush(Color.LightGray, LineType.Solid, Color.Gray)
-            };
-            resolutionHeightTextbox.TextChanged += ResolutionHeightTextbox_TextChanged;
-            resolutionStack.Controls.Add(resolutionHeightTextbox);
-
-            Label pxLabel = new Label(manager)
-            {
-                Text = Languages.OctoClient.Pixels
-            };
-            resolutionStack.Controls.Add(pxLabel);
+            optionsPage.Controls.Add(optionsOptions);
 
             #endregion
 
@@ -199,48 +57,12 @@ namespace OctoAwesome.Client.Screens
             TabPage bindingsPage = new TabPage(manager, Languages.OctoClient.KeyBindings);
             tabs.Pages.Add(bindingsPage);
 
-            ScrollContainer bindingsScroll = new ScrollContainer(manager);
-            bindingsPage.Controls.Add(bindingsScroll);
-
-            StackPanel bindingsStack = new StackPanel(manager)
+            BindingsOptionControl bindingsOptions = new BindingsOptionControl(manager)
             {
-                Orientation = Orientation.Vertical,
-                Padding = new Border(20, 20, 20, 20),
-                Width = 650
+                HorizontalAlignment = HorizontalAlignment.Stretch,
+                VerticalAlignment = VerticalAlignment.Stretch,
             };
-            bindingsScroll.Content = bindingsStack;
-
-            //////////////////////////////KeyBindings////////////////////////////////////////////
-            var bindings = game.KeyMapper.GetBindings();
-            foreach (var binding in bindings)
-            {
-                StackPanel bindingStack = new StackPanel(manager)
-                {
-                    Orientation = Orientation.Horizontal,
-                    HorizontalAlignment = HorizontalAlignment.Stretch,
-                    Height = 35
-                };
-
-                Label lbl = new Label(manager)
-                {
-                    Text = binding.DisplayName,
-                    Width = 480
-                };
-
-                Label bindingKeyLabel = new Label(manager)
-                {
-                    Text = binding.Keys.First().ToString(),
-                    HorizontalAlignment = HorizontalAlignment.Right,
-                    Width = 90,
-                    Background = new BorderBrush(Color.LightGray, LineType.Solid, Color.Gray),
-                    Tag = new object[] { binding.Id, binding.Keys.First() }
-                };
-                bindingKeyLabel.LeftMouseClick += BindingKeyLabel_LeftMouseClick;
-
-                bindingStack.Controls.Add(lbl);
-                bindingStack.Controls.Add(bindingKeyLabel);
-                bindingsStack.Controls.Add(bindingStack);
-            }
+            bindingsPage.Controls.Add(bindingsOptions);
 
             #endregion
 
@@ -269,83 +91,10 @@ namespace OctoAwesome.Client.Screens
             Controls.Add(exitButton);
         }
 
-        private void ResolutionWidthTextbox_TextChanged(Control sender, PropertyEventArgs<string> args)
+        public void NeedRestart()
         {
-            settings.Set("Width", args.NewValue);
-
             exitButton.Visible = true;
             exitButton.Enabled = true;
-        }
-
-        private void ResolutionHeightTextbox_TextChanged(Control sender, PropertyEventArgs<string> args)
-        {
-            settings.Set("Height", args.NewValue);
-
-            exitButton.Visible = true;
-            exitButton.Enabled = true;
-        }
-
-        private void BindingKeyLabel_LeftMouseClick(Control sender, MouseEventArgs args)
-        {
-            object[] data = (object[])sender.Tag;
-            string id = (string)data[0];
-            Keys oldKey = (Keys)data[1];
-
-            Label lbl = (Label)sender;
-
-            MessageScreen screen = new MessageScreen((ScreenComponent)Manager, "Press key...", "", "Cancel");
-            screen.KeyDown += (s, a) =>
-            {
-                game.KeyMapper.RemoveKey(id, oldKey);
-                game.KeyMapper.AddKey(id, a.Key);
-                data[1] = a.Key;
-                settings.Set("KeyMapper-" + id, a.Key.ToString());
-                lbl.Text = a.Key.ToString();
-                Manager.NavigateBack();
-            };
-            Manager.NavigateToScreen(screen);
-        }
-
-        private void SetViewrange(int newRange)
-        {
-            rangeTitle.Text = Languages.OctoClient.Viewrange + ": " + newRange;
-
-            settings.Set("Viewrange", newRange);
-
-            exitButton.Visible = true;
-            exitButton.Enabled = true;
-        }
-
-        private void ChangePath()
-        {
-            System.Windows.Forms.FolderBrowserDialog folderBrowser = new System.Windows.Forms.FolderBrowserDialog();
-            folderBrowser.SelectedPath = settings.Get<string>("ChunkRoot");
-
-            if (folderBrowser.ShowDialog() == System.Windows.Forms.DialogResult.OK)
-            {
-                string path = folderBrowser.SelectedPath;
-                settings.Set("ChunkRoot", path);
-                mapPath.Text = path;
-
-                exitButton.Visible = true;
-                exitButton.Enabled = true;
-            }
-        }
-
-        private void SetPersistence(bool state)
-        {
-            settings.Set("DisablePersistence", state);
-
-            exitButton.Visible = true;
-            exitButton.Enabled = true;
-        }
-
-        private void SetFullscreen(bool state)
-        {
-            settings.Set("EnableFullscreen", state);
-
-            exitButton.Visible = true;
-            exitButton.Enabled = true;
-        }
+        }        
     }
 }
